Break asset library into three UIs
