+++ spec/suitec/asset_library_likes_spec.rb
@@ -21,7 +21,7 @@ describe 'Asset', order: :defined do
     @driver = Utils.launch_browser
     @canvas = Page::CanvasPage.new @driver
     @cal_net = Page::CalNetPage.new @driver
-    @asset_library = Page::SuiteCPages::AssetLibraryPage.new @driver
+    @asset_library = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
     @engagement_index = Page::SuiteCPages::EngagementIndexPage.new @driver
 
     @canvas.log_in(@cal_net, (event.actor = admin).username, Utils.super_admin_password)
@@ -41,7 +41,7 @@ describe 'Asset', order: :defined do
     @admirer_score = @engagement_index.user_score(@driver, @engagement_index_url, @asset_admirer, event)
   end
 
-  after(:all) { @driver.quit }
+  after(:all) { Utils.quit_browser @driver }
 
   describe 'likes' do
