+++ spec/suitec/canvas_discussions_spec.rb
@@ -28,7 +28,7 @@ describe 'A Canvas discussion', order: :defined do
     @canvas.create_generic_course_site(@driver, Utils.canvas_qa_sub_account, @course, [@user_1, @user_2], test_id, tools)
     @engagement_index_url = @canvas.click_tool_link(@driver, LtiTools::ENGAGEMENT_INDEX)
     unless course_id.nil?
-      @asset_library = Page::SuiteCPages::AssetLibraryPage.new @driver
+      @asset_library = Page::SuiteCPages::AssetLibraryListViewPage.new @driver
       @asset_library_url = @canvas.click_tool_link(@driver, LtiTools::ASSET_LIBRARY)
       @asset_library.ensure_canvas_sync(@driver, @asset_library_url)
     end
