+++ spec/suitec/impact_studio_activity_network_spec.rb
@@ -31,7 +31,7 @@ describe 'Impact Studio Activity Network', :order => :defined do
     @driver = Utils.launch_browser
     @canvas = Page::CanvasAnnounceDiscussPage.new @driver
     @cal_net = Page::CalNetPage.new @driver
-    @asset_library_page = Page::SuiteCPages::AssetLibraryPage.new @driver
+    @asset_library_page = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
     @impact_studio_page = Page::SuiteCPages::ImpactStudioPage.new @driver
     @whiteboards_page = Page::SuiteCPages::WhiteboardsPage.new @driver
     @engagement_index_page = Page::SuiteCPages::EngagementIndexPage.new @driver
