+++ spec/suitec/impact_studio_assets_spec.rb
@@ -51,7 +51,7 @@ describe 'The Impact Studio', order: :defined do
     @driver = Utils.launch_browser
     @canvas = Page::CanvasPage.new @driver
     @cal_net = Page::CalNetPage.new @driver
-    @asset_library = Page::SuiteCPages::AssetLibraryPage.new @driver
+    @asset_library = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
     @engagement_index = Page::SuiteCPages::EngagementIndexPage.new @driver
     @impact_studio = Page::SuiteCPages::ImpactStudioPage.new @driver
     @whiteboards = Page::SuiteCPages::WhiteboardsPage.new @driver
