+++ spec/suitec/impact_studio_profile_summary_spec.rb
@@ -26,7 +26,7 @@ describe 'Impact Studio', order: :defined do
     @driver = Utils.launch_browser
     @canvas = Page::CanvasPage.new @driver
     @cal_net = Page::CalNetPage.new @driver
-    @asset_library = Page::SuiteCPages::AssetLibraryPage.new @driver
+    @asset_library = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
     @impact_studio = Page::SuiteCPages::ImpactStudioPage.new @driver
     @engagement_index = Page::SuiteCPages::EngagementIndexPage.new @driver
