+++ spec/suitec/impact_studio_visualizations_spec.rb
@@ -37,7 +37,7 @@ describe 'The Impact Studio', order: :defined do
     @driver = Utils.launch_browser
     @canvas = Page::CanvasAnnounceDiscussPage.new @driver
     @cal_net = Page::CalNetPage.new @driver
-    @asset_library = Page::SuiteCPages::AssetLibraryPage.new @driver
+    @asset_library = Page::SuiteCPages::AssetLibraryDetailPage.new @driver
     @engagement_index = Page::SuiteCPages::EngagementIndexPage.new @driver
     @impact_studio = Page::SuiteCPages::ImpactStudioPage.new @driver
     @whiteboards = Page::SuiteCPages::WhiteboardsPage.new @driver
