+++ util/spec_helper.rb
@@ -76,7 +76,9 @@ require_relative '../pages/boac/student_page'
 require_relative '../pages/boac/teams_list_page'
 require_relative '../pages/boac/api_user_analytics_page'
 require_relative '../pages/suitec/suite_c_pages'
-require_relative '../pages/suitec/asset_library_page'
+require_relative '../pages/suitec/asset_library_list_view_page'
+require_relative '../pages/suitec/asset_library_detail_page'
+require_relative '../pages/suitec/asset_library_manage_assets_page'
 require_relative '../pages/suitec/engagement_index_page'
 require_relative '../pages/suitec/whiteboards_page'
 require_relative '../pages/suitec/impact_studio_page'
